requirements.txt
openai
python-dotenv
pandas
tiktoken
scipy